feat[#135]: 토큰 기반(JWT) 인증방식 로직과 확장성 있게 통합될 수 있도록 구조 개선하기

+++ src/main/java/com/ktb/howard/ktb_community_server/auth/service/SessionAuthService.java
@@ -2,26 +2,33 @@ package com.ktb.howard.ktb_community_server.auth.service;
 
 import com.ktb.howard.ktb_community_server.auth.domain.MemberRole;
 import com.ktb.howard.ktb_community_server.auth.domain.Session;
+import com.ktb.howard.ktb_community_server.auth.dto.AuthResponseDto;
+import com.ktb.howard.ktb_community_server.auth.dto.SessionResponseDto;
 import com.ktb.howard.ktb_community_server.auth.repository.SessionRepository;
 import com.ktb.howard.ktb_community_server.member.domain.Member;
 import com.ktb.howard.ktb_community_server.member.exception.MemberNotFoundException;
 import com.ktb.howard.ktb_community_server.member.exception.PasswordNotMatchedException;
 import com.ktb.howard.ktb_community_server.member.repository.MemberRepository;
+import jakarta.servlet.http.Cookie;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.redis.core.RedisTemplate;
+import org.springframework.http.ResponseCookie;
 import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.Arrays;
 import java.util.Optional;
 import java.util.UUID;
 import java.util.concurrent.TimeUnit;
 
+import static com.ktb.howard.ktb_community_server.filter.SessionAuthFilter.SESSION_COOKIE_NAME;
+
 @Slf4j
 @RequiredArgsConstructor
-@Service
-public class SessionService {
+public class SessionAuthService implements AuthService {
 
     private final MemberRepository memberRepository;
     private final SessionRepository sessionRepository;
@@ -29,7 +36,7 @@ public class SessionService {
     private final PasswordEncoder passwordEncoder;
 
     @Transactional(readOnly = true)
-    public Session login(String email, String password) {
+    public AuthResponseDto login(String email, String password) {
         // 1. 유효한 회원인지 확인
         Member member = memberRepository.findByEmail(email)
                 .orElseThrow(() -> new MemberNotFoundException("입력한 이메일 또는 비밀번호를 확인해주세요."));
@@ -44,38 +51,71 @@ public class SessionService {
                 .sessionId(newSessionId)
                 .memberId(member.getId())
                 .email(member.getEmail())
-                .nickname(member.getNickname())
                 .role(MemberRole.USER)
                 .build();
         sessionRepository.save(session);
-        return session;
+        return SessionResponseDto.builder()
+                .sessionId(newSessionId)
+                .memberId(member.getId())
+                .email(member.getEmail())
+                .nickname(member.getNickname())
+                .role(MemberRole.USER)
+                .build();
     }
 
-    public Optional<Session> findSessionAndSlide(String sessionId) {
-        // 1. CrudRepository를 통해 HGETALL 실행 (세션 데이터 조회 및 유효성 확인)
+    @Override
+    public Optional<AuthResponseDto> refresh(String sessionId) {
         Optional<Session> sessionOpt = sessionRepository.findById(sessionId);
         if (sessionOpt.isPresent()) {
             // 2. RedisTemplate을 통해 EXPIRE 명령어 실행 (슬라이딩)
             String redisKey = Session.KEY_PREFIX + ":" + sessionId;
             redisTemplate.expire(redisKey, 3_600, TimeUnit.SECONDS);
+            Session session = sessionOpt.get();
+            return Optional.of(
+                    SessionResponseDto.builder()
+                            .sessionId(session.getSessionId())
+                            .memberId(session.getMemberId())
+                            .email(session.getEmail())
+                            .role(session.getRole())
+                            .build()
+            );
         }
-        return sessionOpt;
-    }
-
-    // 세션 정보를 업데이트 함
-    public void updateSession(Session currentSession, Member updatedMember) {
-        currentSession.setNickname(updatedMember.getNickname());
-        sessionRepository.save(currentSession);
+        return Optional.empty();
     }
 
     // 로그아웃 수행
-    public void logout(String sessionId) {
+    @Override
+    public void logout(HttpServletRequest request, HttpServletResponse response) {
+        // 헤더를 보고 세션ID 획득 후 세션 스토어에서 세션정보 삭제
+        Optional<String> sessionIdOpt = getSessionId(request);
+        if (sessionIdOpt.isEmpty()) return;
+        String sessionId = sessionIdOpt.get();
         sessionRepository.deleteById(sessionId);
+        // 세션 관련 쿠키 무효화
+        ResponseCookie cookie = ResponseCookie.from(SESSION_COOKIE_NAME, null)
+                .maxAge(0)
+                .path("/")
+                .httpOnly(true)
+                .secure(true)
+                .sameSite("Lax")
+                .build();
+        response.addHeader("Set-Cookie", cookie.toString());
     }
 
     // 특정 사용자와 매칭되는 비밀번호인지 검증
     private boolean checkPassword(Member member, String rawPassword) {
         return passwordEncoder.matches(rawPassword, member.getPassword());
+    }
+
+    private Optional<String> getSessionId(HttpServletRequest request) {
+        Cookie[] cookies = request.getCookies();
+        if (cookies == null) {
+            return Optional.empty();
+        }
+        return Arrays.stream(cookies)
+                .filter(cookie -> cookie.getName().equals(SESSION_COOKIE_NAME))
+                .map(Cookie::getValue) // 쿠키의 값(세션 ID)을 추출
+                .findFirst();
     }
 
 }
